% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output.R
\name{as_tbl_obs}
\alias{as_tbl_obs}
\alias{as_tbl_obs.tbl_df}
\alias{as_tbl_obs.data.frame}
\alias{print.tbl_obs}
\alias{knit_print.tbl_obs}
\title{tibble class for observations}
\usage{
as_tbl_obs(x)

\method{as_tbl_obs}{tbl_df}(x)

\method{as_tbl_obs}{data.frame}(x)

\method{print}{tbl_obs}(x, ...)

\method{knit_print}{tbl_obs}(x, ...)
}
\arguments{
\item{x}{a data frame}
}
\value{
tbl_obs

The constructor makes any data.frame or tibble class tbl_obs.
The provided print method randomly selects a small number of
observations and puts it into a knitr::kable, which looks nice
on both, console and knitr output.
}
\description{
A class for tibbles with observations providing a compact output
}
\author{
Martin Schmettow
}
