% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_manipulation.R
\name{discard_redundant}
\alias{discard_redundant}
\title{Removing redundant variables}
\usage{
discard_redundant(D)
}
\arguments{
\item{D}{data frame}
}
\value{
data frame
}
\description{
all variables that do not vary are discarded
}
\author{
Martin Schmettow
}
