% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math.R
\name{exp_law}
\alias{exp_law}
\title{exp_law}
\usage{
exp_law(pexp, rate, asym, N)
}
\arguments{
\item{pexp}{virtual previous experience}

\item{rate}{rate of learning [-1;1]}

\item{asym}{asymptote of maximum performance}

\item{N}{number of sessions}
}
\value{
numerical
}
\description{
the exponential law of learning in previous experience parametrization
}
\author{
Martin Schmettow
}
