% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_manipulation.R
\name{expand_grid}
\alias{expand_grid}
\title{dplyrified expand.grid}
\usage{
expand_grid(...)
}
\arguments{
\item{...}{factors}
}
\value{
data_frame
}
\description{
works like expand.grid, but returns a tibble data_frame
}
\author{
Martin Schmettow
}
