% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_manipulation.R
\name{left_union}
\alias{left_union}
\title{adding columns that don't exist yet}
\usage{
left_union(df1, df2)
}
\arguments{
\item{df1}{master data frame, where columns are kept}

\item{df2}{data frame where additional columns are taken from}
}
\value{
data frame
}
\description{
adding columns that don't exist yet
}
\author{
Martin Schmettow
}
