% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math.R
\name{mode}
\alias{mode}
\title{mode}
\usage{
mode(x)
}
\arguments{
\item{x}{data}
}
\value{
numerical
}
\description{
estimation of the mode. Uses mofdeest::shorth
Will be replaced by an own implementation to work around the
odd dependencies of modeest (genefilter fro BioConductor).
}
\author{
Martin Schmettow
}
