% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_manipulation.R
\name{reorder_levels}
\alias{reorder_levels}
\title{Re-order levels of a factor}
\usage{
reorder_levels(x, positions)
}
\arguments{
\item{x}{factor}

\item{positions}{vector of positions}
}
\value{
factor
}
\description{
Re-orders the levels of a factor by a vector of new positions.
}
\author{
Martin Schmettow
}
