% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow.R
\name{resave}
\alias{resave}
\title{appending objects to an Rdata file}
\usage{
resave(..., list = character(), file)
}
\arguments{
\item{...}{factors}
}
\value{
data_frame
}
\description{
works like save(), but adds objects, instead of overwriting
This function is slow for large objects. For discussion see
https://stackoverflow.com/questions/11813096/updating-an-existing-rdata-file
}
\author{
Martin Schmettow
}
