% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_manipulation.R
\name{rescale_centered}
\alias{rescale_centered}
\title{centered rescaling}
\usage{
rescale_centered(x, scale = 0.999)
}
\arguments{
\item{x}{numerical vector}

\item{scale}{rescale factor}
}
\value{
numerical vector
}
\description{
symmetrically scales a variable up or down, maintaining the center of the scale
}
\author{
Martin Schmettow
}
