% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_manipulation.R
\name{rescale_zero_one}
\alias{rescale_zero_one}
\title{rescaling to unit interval}
\usage{
rescale_zero_one(x, lower = min(x, na.rm = T), upper = max(x, na.rm =
  T), scale = 1)
}
\arguments{
\item{x}{numerical vector}

\item{lower}{lowest possible value}

\item{upper}{highest possible value}
}
\value{
numerical vector
}
\description{
scales a variable to the interval [0;1]
}
\author{
Martin Schmettow
}
