% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{rrating_scale}
\alias{rrating_scale}
\title{simulate rating scale data}
\usage{
rrating_scale(N, mu, sd, ends = c(1, 5), bin = TRUE)
}
\arguments{
\item{N}{number of simulations}

\item{mu}{mean of underlying normal}

\item{sd}{sd of underlying normal}

\item{ends}{vector of two endpoints c(0,5)}

\item{bin}{whether to bin to integer values}
}
\value{
numerical

Simulates rating scale responses using rlogitnorm.
Note that mu and sd are for the underlying normal distribution.
}
\description{
using a scaled and binned logitnorm distribution
}
\author{
Martin Schmettow
}
