% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_manipulation.R
\name{update_by}
\alias{update_by}
\title{update columns conditionally}
\usage{
update_by(D, by, ...)
}
\arguments{
\item{D}{data frame}

\item{...}{expressions}

\item{filter}{predicate function (like dplyr::filter)}
}
\value{
data frame

Applies mutations to the filtered group, only.
}
\description{
update columns conditionally
}
\examples{
D <- tribble(~group, ~value,  1, 4, 1, 9, 2, -4, 2, -9)

D \%>\% mutate(value = if_else(group == 1, sqrt(value), value))
## Produces NaNs, because sqrt() is evaluated before selection

D \%>\% mutate_by(group == 1, value = sqrt(value))
## sqrt() is only evaluated

}
\author{
Martin Schmettow
}
