% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_manipulation.R
\name{z_score}
\alias{z_score}
\title{Adding z-scores}
\usage{
z_score(D, ...)
}
\arguments{
\item{D}{data frame}

\item{...}{dplyr::select range of variables}
}
\value{
data frame
}
\description{
The (simple) range of columns are z-transformed and added to the data frame (foo_z)
}
\author{
Martin Schmettow
}
