% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_manipulation.R
\name{discard_redundant}
\alias{discard_redundant}
\alias{discard_redundant.default}
\title{Remove redundant variables}
\usage{
discard_redundant(D, except, ...)

\method{discard_redundant}{default}(D, except = c())
}
\arguments{
\item{D}{data frame}

\item{except}{vector of column names to keep}
}
\value{
data frame
}
\description{
all variables that have a constant value are removed from a data frame
}
\author{
Martin Schmettow
}
